+++ SourceCode/Database/dbo/Tables/Stations/ExternalStationCustomer.sql
@@ -5,8 +5,16 @@
     [OpenedYear]        SMALLINT      NULL,
     [ClosedYear]        SMALLINT      NULL,
     CONSTRAINT [PK_ExternalStationCustomer] PRIMARY KEY CLUSTERED ([Id] ASC),
-    CONSTRAINT [FK_ExternalStationCustomer_ExternalStation] FOREIGN KEY ([ExternalStationId]) REFERENCES [dbo].[ExternalStation] ([Id]) ON DELETE CASCADE
+    CONSTRAINT [FK_ExternalStationCustomer_ExternalStation] FOREIGN KEY ([ExternalStationId]) REFERENCES [dbo].[ExternalStation] ([Id]) 
 );
+
+GO
+CREATE TRIGGER [dbo].[DeleteExternalStationCustomer] ON [ExternalStationCustomer] INSTEAD OF DELETE
+AS
+BEGIN
+    DELETE FROM [ExternalStationCustomerCargo] WHERE ExternalStationCustomerId IN (SELECT [Id] FROM DELETED)
+    DELETE FROM [ExternalStationCustomer] WHERE Id IN (SELECT [Id] FROM DELETED)
+END