+++ SourceCode/Database/dbo/Tables/Stations/ExternalStationCustomerCargo.sql
@@ -16,8 +16,9 @@
     CONSTRAINT [FK_ExternalStationCustomerCargo_CargoDirection] FOREIGN KEY ([DirectionId]) REFERENCES [dbo].[CargoDirection] ([Id]),
     CONSTRAINT [FK_ExternalStationCustomerCargo_CargoUnit] FOREIGN KEY ([QuantityUnitId]) REFERENCES [dbo].[CargoUnit] ([Id]),
     CONSTRAINT [FK_ExternalStationCustomerCargo_OperatingDay] FOREIGN KEY ([OperatingDayId]) REFERENCES [dbo].[OperatingDay] ([Id]),
-    CONSTRAINT [FK_ExternalStationCustomerCargo_StationCustomer] FOREIGN KEY ([ExternalStationCustomerId]) REFERENCES [dbo].[ExternalStationCustomer] ([Id]) 
+    CONSTRAINT [FK_ExternalStationCustomerCargo_StationCustomer] FOREIGN KEY ([ExternalStationCustomerId]) REFERENCES [dbo].[ExternalStationCustomer] ([Id])
 );
+
 GO
 CREATE TRIGGER [DeleteExternalStationCustomerCargo] ON [ExternalStationCustomerCargo] INSTEAD OF DELETE 
 AS
@@ -25,6 +26,7 @@ BEGIN
     DELETE FROM [StationCustomerWaybill] WHERE ISNULL([OtherExternalCustomerCargoId],0) IN (SELECT [Id] FROM DELETED)
     DELETE FROM [ExternalStationCustomerCargo] WHERE [Id] IN (SELECT [Id] FROM DELETED)
 END
+
 GO
 CREATE NONCLUSTERED INDEX [IX_ExternalStationCustomerCargo_StationCustomerId]
     ON [dbo].[ExternalStationCustomerCargo] ([ExternalStationCustomerId] ASC)