+++ SourceCode/Services/Extensions/ModuleExtensions.cs
@@ -20,6 +20,11 @@ static public class ModuleExtensions
         module is null || module.Station is null ? string.Empty :
         $"{module.Station.FullName}: {module.Station.Signature}";
 
+    public static string NameOrStationNameWithSignature(this Module? module) =>
+        module is null ? string.Empty :
+        module.Station is null ? module.FullName :
+        module.StationNameWithSignature();
+
     public static bool IsGroupOwned([NotNullWhen(true)] this Module? module) =>
         module is not null && module.ModuleOwnerships is not null && module.ModuleOwnerships.Any(mo => mo.GroupId > 0);
     public static bool IsPersonOwned([NotNullWhen(true)] this Module? module) =>